Send emails directly over SMTP
This adds gsettings for SMTP server/port etc. but adds no UI for those.
The default is localhost:25, no auth, no TLS.

There's some code for SMTP authentication, but it is incomplete (no
keyring integration, no way to provide a password).

The code is known not to work under Python 2 (smtplib.SMTP is not a
context manager).

Some minor refactorings snuck in.